move secrets to utils folder
